requirements.txt
torch
transformers
numpy
scikit-learn
opencv-python
Pillow
joblib
gradio
pyngrok
matplotlib
seaborn
jupyter